+++ custom_python/add_io_to_python.cpp
@@ -52,8 +52,8 @@ void GidPostIO_WriteMesh( TGidPostIOType& dummy, typename TGidPostIOType::MeshTy
 }
 
 template<typename TGidPostIOType>
-void PrintElementalPartitionIndex( TGidPostIOType& dummy, const Variable<double>& rVariable,
-                                   typename TGidPostIOType::ModelPartType& r_model_part, double SolutionTag, int rank)
+void PrintElementalPartitionIndex( TGidPostIOType& dummy, const Variable<int>& rVariable,
+                                   typename TGidPostIOType::ModelPartType& r_model_part, double SolutionTag, int rank )
 {
     dummy.PrintElementalPartitionIndex( rVariable, r_model_part, SolutionTag, rank );
 }
